requirements.txt
-i https://pypi.org/simple
asgiref==3.7.2; python_version >= '3.7'
backports.zoneinfo==0.2.1; python_version < '3.9'
certifi==2023.7.22; python_version >= '3.6'
charset-normalizer==3.2.0; python_full_version >= '3.7.0'
coreapi==2.3.3
coreschema==0.0.4
django==4.2.5
djangorestframework==3.14.0
idna==3.4; python_version >= '3.5'
install==1.3.5
itypes==1.2.0
jinja2==3.1.2; python_version >= '3.7'
markupsafe==2.1.3; python_version >= '3.7'
pillow==10.0.1
pip==23.2.1
pytz==2023.3.post1
requests==2.31.0; python_version >= '3.7'
sqlparse==0.4.4; python_version >= '3.5'
typing-extensions==4.8.0; python_version < '3.11'
tzdata==2023.3; sys_platform == 'win32'
uritemplate==4.1.1; python_version >= '3.6'
urllib3==2.0.4; python_version >= '3.7'
